requirements.txt
# Smart Factory Defect Detection System v2.0
# Python 3.8+ 필요

# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# CLIP Model
open-clip-torch>=2.20.0

# Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0
scikit-image>=0.21.0

# Scientific Computing
numpy>=1.24.0
scipy>=1.11.0

# Visualization
matplotlib>=3.7.0

# Optional: SAM (Segment Anything Model) - 고급 객체 분할
# segment-anything>=1.0
# git+https://github.com/facebookresearch/segment-anything.git

# Optional: Performance Monitoring
# tqdm>=4.65.0  # 진행률 표시바
# psutil>=5.9.0  # 메모리/CPU 모니터링

# Development Tools (선택사항)
# pytest>=7.4.0
# black>=23.7.0
# flake8>=6.1.0